Alguns gráficos do dashboard

+++ Desapegando/src/Desapegando.Application/ViewModels/DashboardViewModel.cs
@@ -9,9 +9,9 @@
         public IEnumerable<NovosCondominos7DiasViewModel> NovosCondominos7Dias { get; set; }
         public IEnumerable<Vendas7DiasViewModel> Vendas7DiasViewModel { get; set; }
 
-        public int TotalProdutosVendidosUltimos7Dias { get; set; }
-        public int TotalProdutosDisponiveisUltimos7Dias { get; set; }
-        public int TotalProdutosDesistidosUltimos7Dias { get; set; }
+        public decimal TotalProdutosVendidosUltimos7Dias { get; set; }
+        public decimal TotalProdutosDisponiveisUltimos7Dias { get; set; }
+        public decimal TotalProdutosDesistidosUltimos7Dias { get; set; }
 
         public int NovasCampanhasDisponiveisUlitmos30Dias { get; set; }
     }
